+++ Project/MusicDetailViewKnt.m
@@ -978,10 +978,10 @@ static inline void MusicDetailViewKntBuildDifficultyButton(MusicDetailViewKnt *s
     [self.buttonLink setHidden:(info.iTunesURL == nil)];
     [self.btnRecommendTwitter setHidden:NO];
     [self.btnRecommendFacebook setHidden:NO];
-    [levelNumView[0] setImage:levelNumImg[self.levelBas]];
-    [levelNumView[1] setImage:levelNumImg[self.levelAdv]];
-    [levelNumView[2] setImage:levelNumImg[self.levelExt]];
-    [levelNumView[kExtendLevelNumIndex] setImage:levelNumImg[self.levelExt]];
+    [levelNumView[0] setImage:levelNumImg[(int)self.levelBas]];
+    [levelNumView[1] setImage:levelNumImg[(int)self.levelAdv]];
+    [levelNumView[2] setImage:levelNumImg[(int)self.levelExt]];
+    [levelNumView[kExtendLevelNumIndex] setImage:levelNumImg[(int)self.levelExt]];
     [levelNumView[kExtendLevelNumIndex] setAlpha:0.0];
     [self resetScore];
     [self putScore:score];
@@ -1724,7 +1724,7 @@ static inline void MusicDetailViewKntBuildDifficultyButton(MusicDetailViewKnt *s
     [editTxt[1] setText:editorInfo[@"editorName"]];
     [editTxt[2] setText:editorInfo[@"comment"]];
     int level = [editorInfo[@"level"] intValue];
-    [levelNumView[kExtendLevelNumIndex] setImage:levelNumImg[level]];
+    [levelNumView[kExtendLevelNumIndex] setImage:levelNumImg[(int)level]];
     [levelNumView[kExtendLevelNumIndex] setAlpha:1.0];
     [self.controller dismissViewControllerAnimated:YES completion:nil];
     [self.controller enableCoverTap];
@@ -1898,7 +1898,7 @@ static inline void MusicDetailViewKntBuildDifficultyButton(MusicDetailViewKnt *s
         if (isExtend && (self.extendInfo.extendFlag & (1 << i)) != 0) {
             level = extendLevels[i];
         }
-        [levelNumView[i] setImage:levelNumImg[level]];
+        [levelNumView[i] setImage:levelNumImg[(int)level]];
     }
 
     // The current difficulty's hold mark drives the app hold flag (except on the edit page).
@@ -2138,7 +2138,7 @@ static inline void MusicDetailViewKntBuildDifficultyButton(MusicDetailViewKnt *s
     [editTxt[1] setText:editorInfo[@"editorName"]];
     [editTxt[2] setText:editorInfo[@"comment"]];
     int level = [editorInfo[@"level"] intValue];
-    [levelNumView[kExtendLevelNumIndex] setImage:levelNumImg[level]];
+    [levelNumView[kExtendLevelNumIndex] setImage:levelNumImg[(int)level]];
     [levelNumView[kExtendLevelNumIndex] setAlpha:1.0];
     if (dlFlag == kEditDownloadFlag) {
         [infoBtn setAdjustsImageWhenHighlighted:NO];
